\name{pfda-package}
\alias{pfda-package}
\docType{package}
\title{Paired Functional Data Analysis, using Principle Components}
\description{Paired functional data analysis using functional principle components.  The principal components are formed using orthogonalized spline basis.
}
\details{
	Computes the models of functional principal components for the single, single binary, paired, with mixed binary and continuous responses, and the additive model.  All the models are accessed through the \code{\link{pfda}} function.
  Also of interest is the functions for working with \code{\link[orthogonalsplinebasis]{SplineBasis}}.
}
\references{
\emph{Joint modelling of paired sparse functional data using principal components}
Lan Zhou; Jianhua Z. Huang; Raymond J. Carroll
Biometrika 2008 95: 601-619
}
\author{
	Andrew Redd
	Maintainer: Andrew Redd <aredd at stat.tamu.edu>
}
\keyword{package}
